column schema upsert
batches retry on a missing column, after adding the column. Repeated
retries are very possible - it's not expected that column addition
will be so frequent as to require batch schema resolution but it's
very doable if necessary.

Dimensions and measurement types other than histogram are handled.

Connection management is extracted into a PostgresConnector which
vends a living connection at each interaction. This simplifies state
management in PostgresSink and focuses data sinking therein, with
connection concerns living in the Connector. All interactions are
done under a Transaction right now.