feat: concurrency
Utilizes multiple connections now. Bootstrapping schema is very noisy
still but once it's bootstrapped the data flows pretty seamlessly.

Need to improve batching still so it's not 1:1 api call to db interaction
and might still need to look at how the copy is happening - it's going to
be a tooon of transactions which will put pressure on postgres's xid
wraparound vacuuming.

Batching + widening transactions, and bootstrapping schema in one shot
would all help support a world with 1 txn per huge batch, which is much
more sympathetic to the database.

Also adds more number types and renames them to be plain I32, F64 etc.
Will need to patch the kotlin client; opportunities for renumbering protos
will end once this is actually in use somewhere.