add authorization
simple api key auth. You pass in a list of keys to the server, and
you can configure clients to pass it as well.

When a server is configured with authorization, the requests need
to have an authorization header. They also need the value of that
header to match one of the configured values.